Method call sugar for closure stuff.
Notice the change in closure usage though:

  foo {| someStatement(); }

is equiv to

  foo( {| someStatement(); } )

Also allowed:

  foo(x,y){| someStatement(); }

is equiv to

  foo(x,y, {| someStatement(); })

Normal with params is:

  {|x|x.whatever();}

So, general rules:

-- If you have no parameters, the prefix of {| is required.

-- If you have parameters, the prefix of {| is required, and after
   you formal parameter decls, another | is required to introduce
   the statement block.

So, we can't get this:

  foo.each{ ...block... }

but rather

  foo.each{| ...block... }

"Why oh why?" you might exclaim? Pontificate upon:

	if foo {
		...block...
	}

So here, it actually thinks the block is a sugary parameter on
a method named 'foo' which is being called without parentheses.

Now, with the pipe in the prefix to differentiate, you can
do either construct more deterministically:

	if foo{!
		// calling foo({!derr();}) for conditional
		derrr();
	} {
		// block of if branch
	}
or...

	if foo {
		// block of if branch
	}


git-svn-id: https://svn.codehaus.org/groovy/trunk/groovy/groovy-core@187 a5544e8c-8a19-0410-ba12-f9af4593a198