Fully working BSF engine including variable passing context etc.
Now rather than just Groovy scripts being put into the static main() method, the scripts now derive from Script, the code is in the run() instance method and the Script object can take a ScriptContext to allow variables to be passed into and out of the script.

This makes for easy BSF integration. It will also be useful for interactive shells too; the variables can be shared across script invocations. Indeed each script can be in its own classloader and only the ScriptContext shared.


git-svn-id: https://svn.codehaus.org/groovy/trunk/groovy/groovy-core@251 a5544e8c-8a19-0410-ba12-f9af4593a198