* BIG CHANGE TO TEST PROCEDURES
* In order to speed up the tests, I have grouped the tests into vague groupings
called TestSuites in junit speak.

* If you add any new tests, please add to one of the existing suites,
exactly where doesn't matter, unless it is slightly long running test
in which case put it in UberTestCaseLongRunningTests.java

* The individual test results no longer print to screen during build
(just the number of failures in each suite)

* In order to dig deeper into each failure, look in
target/test-reports/TEST-???.xml  and search for 'failure' text string

This should now improve build time by an order of 5 or so. YMMV


jez


git-svn-id: https://svn.codehaus.org/groovy/trunk/groovy/groovy-core@1540 a5544e8c-8a19-0410-ba12-f9af4593a198